+++ src/minecraft/co/uk/flansmods/client/model/ModelPlane.java
@@ -330,7 +330,7 @@ public class ModelPlane extends ModelDriveable
 	}
 
 	@Override
-	public void translateAll(int x, int y, int z)
+	public void translateAll(float x, float y, float z)
 	{
 		super.translateAll(x, y, z);
 		translate(noseModel, x, y, z);
